[Cleaning House] Delete some old, unused, and untested code (#1166)
