--- guidance/_cpp/__init__.py
@@ -1,1 +1,0 @@
-from .byte_trie import ByteTrie
