+++ guidance/library/__init__.py
@@ -1,5 +1,5 @@
 # import functions that can be called directly
-from ._gen import gen, call_tool, will_gen, regex
+from ._gen import gen, call_tool, regex
 from ._image import image, gen_image
 from ._audio import audio, gen_audio
 from ._video import video, gen_video
@@ -14,16 +14,11 @@ from .._grammar import token_limit
 # context blocks
 from ._block import block
 from ._role import role, system, assistant, user #, function, instruction, indent_roles
-from ._format import monospace
-from ._silent import silent
-from ._set_var import set_var
-from ._set_attribute import set_attribute
 
 # from ..models._model import context_free
 
 # stateless library functions
 from ._sequences import one_or_more, zero_or_more, at_most_n_repeats, exactly_n_repeats, sequence
-from ._prefix_tree import prefix_tree
 from ._substring import substring
 from ._optional import optional
 from ._tool import Tool
