+++ guidance/library/_gen.py
@@ -1,10 +1,8 @@
 from typing import Union, Optional
-import regex as regex_module
 import logging
 from .._guidance import guidance
 from .._grammar import regex as regex_node, select, quote_regex, capture, with_temperature, gen as grammar_gen
 from ._block import block
-from ._silent import silent
 from ._tool import Tool
 
 logger = logging.getLogger(__name__)
@@ -169,83 +167,6 @@ def gen(
         temperature=temperature,
         max_tokens=max_tokens
     )
-
-
-def click_loop_start(id, total_count, echo, color):
-    click_script = (
-        """
-function cycle_IDVAL(button_el) {
-var i = 0;
-while (i < 50) {
-var el = document.getElementById("IDVAL_" + i);
-if (el.style.display == "inline") {
-    el.style.display = "none";
-    var next_el = document.getElementById("IDVAL_" + (i+1));
-    if (!next_el) {
-        next_el = document.getElementById("IDVAL_0");
-    }
-    if (next_el) {
-        next_el.style.display = "inline";
-    }
-    break;
-}
-i += 1;
-}
-button_el.innerHTML = (((i+1) % TOTALCOUNT) + 1)  + "/" + TOTALCOUNT;
-}
-cycle_IDVAL(this);""".replace(
-            "IDVAL", id
-        )
-        .replace("TOTALCOUNT", str(total_count))
-        .replace("\n", "")
-    )
-    out = f"""<div style='background: rgba(255, 255, 255, 0.0); border-radius: 4px 0px 0px 4px; border: 1px solid {color}; border-right: 0px; padding-left: 3px; padding-right: 3px; user-select: none; color: {color}; display: inline; font-weight: normal; cursor: pointer' onClick='{click_script}'>1/{total_count}</div>"""
-    out += f"<div style='display: inline;' id='{id}_0'>"
-    return "<||_html:" + out + "_||>"
-
-
-def click_loop_mid(id, index, echo):
-    alpha = 1.0 if not echo else 0.5
-    out = f"</div><div style='display: none; opacity: {alpha}' id='{id}_{index}'>"
-    return "<||_html:" + out + "_||>"
-
-
-@guidance
-def gen_line(lm, *args, **kwargs):
-    return lm.gen(*args, suffix="\n", **kwargs)
-
-
-@guidance
-def gen_quote(lm, name=None, quote='"', *args, **kwargs):
-    return lm(quote).gen(*args, name=name, suffix=quote, **kwargs)
-
-
-@guidance
-def will_gen(lm, stop=None, stop_regex=None, ignore_spaces=False, max_tokens=30):
-    # this is obviously not the right implementation, just here so we can explore
-    if stop and not isinstance(stop, list):
-        stop = [stop]
-    if stop_regex and not isinstance(stop_regex, list):
-        stop_regex = [stop_regex]
-    assert (stop is not None) or (stop_regex is not None)
-    if not stop:
-        stop = []
-    if not stop_regex:
-        stop_regex = []
-    regexes = [regex_module.escape(x) for x in stop + stop_regex]
-    optional_space = "\\s*" if ignore_spaces else ""
-    pattern = regex_module.compile(f'{optional_space}({"|".join(regexes)})')
-    lm2 = lm
-    with silent():
-        for _ in range(max_tokens):
-            lm2 += gen("temp_variable", list_append=True, max_tokens=1)
-            if not lm2["temp_variable"] or not pattern.match(
-                "".join(lm2["temp_variable"]), partial=True
-            ):
-                return False
-            if pattern.match("".join(lm2["temp_variable"]), partial=False):
-                return True
-    return False
 
 
 @guidance
