--- guidance/library/_set_attribute.py
@@ -1,2 +1,0 @@
-def set_attribute(name, value=True):
-    raise NotImplementedError("set_attribute block not currently supported and may be deprecated")
