--- guidance/library/_silent.py
@@ -1,2 +1,0 @@
-def silent():
-    raise NotImplementedError("silent block not currently supported and may be deprecated")
