+++ tests/model_integration/test_tokenizers.py
@@ -1,15 +1,10 @@
 import pytest
-import tiktoken
-
-from guidance import models
-
 from tests.tokenizer_common import TOKENIZER_ROUND_TRIP_STRINGS, BaseTestTransformerTokenizers
 
 # These are not _strictly_ unit tests, since they refer
 # to specific tokenisers. However, tokenisers are small,
 # so if the tokeniser can be loaded separately from the
-# model (as is the case for transformers and tiktoken),
-# then this is a good place to have them live.
+# model, then this is a good place to have them live.
 
 # The LlamaCpp tokenisers have tests in test_llamacpp.py
 # since those tokenisers cannot be loaded separately from
@@ -38,20 +33,3 @@ class TestUnauthenticatedTransformerTokenizers(BaseTestTransformerTokenizers):
     @pytest.mark.parametrize("target_string", TOKENIZER_ROUND_TRIP_STRINGS)
     def test_string_roundtrip(self, model_name: str, target_string: str):
         self.base_string_roundtrip(model_name, target_string)
-
-
-class TestTiktoken:
-    MODELS = ["gpt-3.5-turbo", "gpt-4"]
-
-    @pytest.mark.parametrize("model_name", MODELS)
-    @pytest.mark.parametrize("target_string", TOKENIZER_ROUND_TRIP_STRINGS)
-    def test_string_roundtrip(self, model_name: str, target_string: str):
-        from guidance.models._grammarless import GrammarlessTokenizer
-        my_tik = tiktoken.encoding_for_model(model_name)
-        my_tok = GrammarlessTokenizer(my_tik)
-
-        encoded = my_tok.encode(target_string.encode())
-        decoded = my_tok.decode(encoded)
-        final_string = decoded.decode()
-
-        assert final_string == target_string
